Add option to exclude alerts from the results
The use case behind this: sometimes you want to define
a so called Watchdog or DeadMansSwitch alert that is always
firing, in order to monitoring that the alerting is working.

When such a Watchdog is defined the list of all alerts will
always be Critical. Thus we add a flag to exclude certain alerts.